chore: css refactoring
