+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-11-11T10:00:55.132Z\n"
-"PO-Revision-Date: 2020-11-11T10:00:55.132Z\n"
+"POT-Creation-Date: 2020-11-11T16:43:59.124Z\n"
+"PO-Revision-Date: 2020-11-11T16:43:59.124Z\n"
 
 msgid "Maps"
 msgstr ""
@@ -541,7 +541,7 @@ msgstr ""
 msgid "More actions"
 msgstr ""
 
-msgid "Data table"
+msgid "Show data table"
 msgstr ""
 
 msgid "Open as chart"
