+++ package.json
@@ -79,7 +79,6 @@
         "react": "^16.13.1",
         "react-color": "^2.18.1",
         "react-dom": "^16.13.1",
-        "react-jss": "^8.6.1",
         "react-redux": "^7.2.1",
         "react-sortable-hoc": "^1.11.0",
         "react-virtualized": "^9.22.2",
