+++ src/components/app/styles/App.module.css
@@ -1,14 +1,15 @@
-body {
+.app {
     font-family: 'Roboto', sans-serif;
 }
+
 /* Scrollbar width */
-::-webkit-scrollbar {
+.app ::-webkit-scrollbar {
     width: 6px;
     height: 6px;
 }
 
 /* Scrollbar handle */
-::-webkit-scrollbar-thumb {
+.app ::-webkit-scrollbar-thumb {
     background: #cbcdcf;
     border-radius: 3px;
 }
