+++ src/components/datatable/styles/BottomPanel.module.css
@@ -1,4 +1,4 @@
-.BottomPanel {
+.bottomPanel {
     position: absolute;
     right: 0;
     bottom: 0;
@@ -7,7 +7,7 @@
     box-shadow: 0 -2px 3px rgba(0, 0, 0, 0.1);
 }
 
-.BottomPanel-close {
+.closeIcon {
     position: absolute;
     top: 3px;
     right: 3px;
