+++ src/components/datatable/styles/ColorCell.module.css
@@ -1,4 +1,4 @@
-.ColorCell {
+.colorCell {
     height: 32px;
     line-height: 32px;
     text-align: center;
