+++ src/components/datatable/styles/ColumnHeader.module.css
@@ -1,8 +1,8 @@
-.ColumnHeader {
+.columnHeader {
     position: relative;
 }
 
-.ColumnHeader-label {
+.label {
     display: inline-block;
     position: absolute;
     top: 7px;
@@ -12,7 +12,7 @@
     overflow: hidden;
 }
 
-.ColumnHeader svg {
+.columnHeader svg {
     position: absolute;
     top: 5px;
     right: 0;
