+++ src/components/download/styles/DownloadLegend.module.css
@@ -1,11 +1,8 @@
 .downloadLegend {
+    composes: box from '../../../styles/map.css';
     position: absolute;
     padding: var(--spacers-dp8) var(--spacers-dp8) 0 var(--spacers-dp8);
     z-index: 998;
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-    border-radius: var(--spacers-dp4);
     font-size: 14px;
     max-width: 220px;
 }
