--- src/components/legend/Legend.css
@@ -1,55 +1,0 @@
-.Legend {
-    padding: 0 16px 16px 32px;
-    margin: 0;
-}
-
-.Legend-description {
-    padding-bottom: 12px;
-}
-
-.Legend-period {
-    padding-bottom: 12px;
-}
-
-.Legend-filters {
-    padding-bottom: 12px;
-}
-
-.Legend-unit {
-    margin-left: 32px;
-    line-height: 24px;
-}
-
-.Legend-source {
-    padding-top: 16px;
-    font-size: 12px;
-}
-
-.Legend table {
-    border-collapse: collapse;
-    border-spacing: 0;
-}
-
-.Legend table tr {
-    height: 24px;
-}
-
-.Legend table th {
-    min-width: 24px;
-    height: 24px;
-    padding: 0;
-    vertical-align: middle;
-}
-
-.Legend table th span {
-    display: inline-block;
-    width: 100%;
-    height: 100%;
-    background-repeat: no-repeat;
-    background-position: center center;
-}
-
-.Legend table td {
-    padding-left: 10px;
-    white-space: nowrap;
-}
