+++ src/components/map/styles/MapName.module.css
@@ -12,15 +12,9 @@
     pointer-events: none;
 }
 
-.mapName div {
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-    border-radius: 4px;
-    margin: var(--spacers-dp8);
-}
-
 .name {
+    composes: box from '../../../styles/map.css';
+    margin: var(--spacers-dp8);
     padding: 6px 8px;
     line-height: 17px;
 }
