+++ src/components/map/styles/PeriodName.module.css
@@ -1,4 +1,4 @@
-.period {
+.periodName {
     display: flex;
     justify-content: center;
     pointer-events: none;
@@ -9,11 +9,8 @@
     z-index: 998;
 }
 
-.period div {
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-    border-radius: 4px;
+.period {
+    composes: box from '../../../styles/map.css';
     padding: 3px 5px;
     font-size: 12px;
 }
