+++ src/components/periods/styles/Timeline.module.css
@@ -1,8 +1,5 @@
 .timeline {
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-    border-radius: 4px;
+    composes: box from '../../../styles/map.css';
     width: calc(100% - 20px);
     position: absolute;
     left: 10px;
