+++ src/components/plugin/styles/Legend.css
