+++ src/components/plugin/styles/MapName.module.css
@@ -9,11 +9,8 @@
     z-index: 998px;
 }
 
-.mapName div {
-    background-color: rgba(255, 255, 255, 0.9);
-    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
-        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
-    border-radius: 4px;
+.name {
+    composes: box from '../../../styles/map.css';
     padding: 5px 8px 3px;
     font-size: 12px;
     text-align: center;
