+++ src/styles/map.css
@@ -1,0 +1,6 @@
+.box {
+    background-color: rgba(255, 255, 255, 0.9);
+    box-shadow: 0px 2px 1px -1px rgba(0, 0, 0, 0.2),
+        0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12);
+    border-radius: 4px;
+}
